add PT thread helper
