+++ hyperdbg/libhyperdbg/header/debugger/misc/pci-id.h
@@ -46,6 +46,7 @@ typedef struct Vendor
 //////////////////////////////////////////////////
 //					  Functions                 //
 //////////////////////////////////////////////////
+
 Vendor *
 GetVendorById(UINT16 VendorId);
 void
