+++ hyperdbg/libhyperdbg/libhyperdbg.vcxproj
@@ -157,6 +157,7 @@ msbuild "$(SolutionDir)dependencies\zydis\msvc\Zydis.sln" /m /p:Configuration="R
     <ClInclude Include="header\debugger\misc\assembler.h" />
     <ClInclude Include="header\debugger\misc\inipp.h" />
     <ClInclude Include="header\debugger\misc\pci-id.h" />
+    <ClInclude Include="header\debugger\misc\pt-helper.h" />
     <ClInclude Include="header\debugger\script-engine\script-engine.h" />
     <ClInclude Include="header\debugger\script-engine\symbol.h" />
     <ClInclude Include="header\debugger\tests\tests.h" />
@@ -231,6 +232,7 @@ msbuild "$(SolutionDir)dependencies\zydis\msvc\Zydis.sln" /m /p:Configuration="R
     <ClCompile Include="code\debugger\misc\callstack.cpp" />
     <ClCompile Include="code\debugger\misc\disassembler.cpp" />
     <ClCompile Include="code\debugger\misc\pci-id.cpp" />
+    <ClCompile Include="code\debugger\misc\pt-helper.cpp" />
     <ClCompile Include="code\debugger\misc\readmem.cpp" />
     <ClCompile Include="code\debugger\script-engine\script-engine-wrapper.cpp" />
     <ClCompile Include="code\debugger\script-engine\script-engine.cpp" />
